[AGILE-296] Inline burndown chart helper into burndown widget
